requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.3.1=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
cachetools=5.5.2=pypi_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
contourpy=1.3.0=pypi_0
cudatoolkit=11.2.72=h2bc3f7f_0
cudnn=8.1.0.77=h90431f1_0
cycler=0.12.1=pypi_0
expat=2.7.1=h6a678d5_0
flatbuffers=25.2.10=pypi_0
fonttools=4.59.2=pypi_0
gast=0.6.0=pypi_0
google-auth=2.40.3=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.75.0=pypi_0
h5py=3.14.0=pypi_0
idna=3.10=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.5.2=pypi_0
joblib=1.5.2=pypi_0
keras=2.8.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.7=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libclang=18.1.1=pypi_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_5
libgcc-ng=15.1.0=h69a702a_5
libgomp=15.1.0=h767d61c_5
libstdcxx-ng=11.2.0=h1234567_1
libxcb=1.17.0=h9b100fa_0
llvmlite=0.38.1=pypi_0
markdown=3.9=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.5.3=pypi_0
ncurses=6.5=h7934f7d_0
numba=0.55.2=pypi_0
numpy=1.21.6=pypi_0
oauthlib=3.3.1=pypi_0
opencv-contrib-python=4.6.0.66=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=3.5.3=h26f9b46_1
opt-einsum=3.4.0=pypi_0
packaging=25.0=pypi_0
pillow=11.3.0=pypi_0
pip=25.2=pyhc872135_0
protobuf=3.19.6=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pydot=4.0.1=pypi_0
pyparsing=3.2.4=pypi_0
python=3.9.23=he99959a_0
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.3=hc2a1206_0
requests=2.32.5=pypi_0
requests-oauthlib=2.0.0=pypi_0
rsa=4.9.1=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.8.1=pypi_0
setuptools=80.9.0=pypi_0
six=1.17.0=pypi_0
sqlite=3.50.2=hb25bd0a_1
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow=2.8.4=pypi_0
tensorflow-estimator=2.8.0=pypi_0
tensorflow-io-gcs-filesystem=0.37.1=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.15=h54e0aa7_0
typing-extensions=4.15.0=pypi_0
tzdata=2025b=h04d1e81_0
urllib3=2.5.0=pypi_0
werkzeug=3.1.3=pypi_0
wget=3.2=pypi_0
wheel=0.45.1=py39h06a4308_0
wrapt=1.17.3=pypi_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zipp=3.23.0=pypi_0
zlib=1.2.13=hd233ad5_2
